backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
mysqlclient==2.2.0
python-decouple==3.8
animeflv==0.3.1
setuptools==80.9.0
PyJWT==2.8.0
cloudscraper==1.2.71
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.4